WIP: Control flow merging (#60)
* start working on control flow merging

* some progress on control flow merging

* add isMerged and a few comments

* implement counting of how many selects are needed for merging, still have to actually update the cost

* add merging cost calc

* some progress on control flow merging; currently working on creating the select instructions for generating the code; we'll want to use one LoopTree Predicate instruction vector to ease all comparisons (not needing to shift indices and vectors around). The primary benefit of keeping the predicates around for a while is to defer decision making about code generating until later, and deciding how to handle things like `div` where exceptions may be thrown from invalid inputs, meaning we cannot actually take a naive 'evaluate both sides' approach. A second thing to consider at the moment is that our cost modeling should look to parents for instructions and consider if we support predicated instructions that are more or less free, e.g. if we're comparing two `*` instructions with operands `(a+x)`,`y` and `x`,`y`, respectively, then we can merge `a+x` and `x` for free on arches supporting predication, such as AVX512 or SVE.

* enable bugprone, start working on addressing it, stronger typing

* some bug fixes

* more minor progress

* more conformance changes; tests pass

* added createSelect

* further progress

* more strong typing...

* use tag instead of CRTP for strong axis int types

* fix reverse iterators

* remap

* compiles and runs locally

* make gcc happy

+++ include/Orthogonalize.hpp
@@ -1,8 +1,8 @@
 #pragma once
-#include "./ArrayReference.hpp"
-#include "./Loops.hpp"
 #include "./Math.hpp"
 #include "./NormalForm.hpp"
+#include "./VectorGreatestCommonDivisor.hpp"
+#include "Rational.hpp"
 #include <cstdint>
 #include <llvm/ADT/SmallVector.h>
 
@@ -14,7 +14,7 @@
     if (A.numRow() == 1)
         return A;
     llvm::SmallVector<Rational, 8> buff;
-    buff.resize_for_overwrite(A.numCol());
+    buff.resize_for_overwrite(size_t(A.numCol()));
     for (size_t i = 1; i < A.numRow(); ++i) {
         for (size_t j = 0; j < A.numCol(); ++j)
             buff[j] = A(i, j);
